ghec: typo fixed in fields name declaration and seconds field moved from Int to Float (v1.1 exposes it)

+++ oq-ui-api/geonode/ghec_viewer/models.py
@@ -1,7 +1,7 @@
 from django.contrib.gis.db import models
 
 # En    Source    Year      Mo        Da        Ho        Mi        Se        Area      Lat       Lon       LatUnc    LonUnc    EpDet     Dep       Io        Msource   M         MUnc      MType     MDet      MDPSource         MDPn      MDPIx     MDPsc     Remarks GEHid
-# en(y) src(y)    yea(y)    mon(n)    day(n)    hou(n)    min(n)    sec(n)    are(n)    lat(y)    lon(y)    latunc(n) lonunc(n) epdet(n)  dep(n)    io(n)     msrc(y)   m(y)      munc(n)   mtyp(n)   mdet(n)   mdpsrc(n)         mdpn(n)   mdpix(n)  mdpsc(n)  rem(n)  gid
+# en(y) src(y)    yea(y)    mon(n)    day(n)    hou(n)    min(n)    sec(n)    are(n)    lat(y)    lon(y)    latunc(n) lonunc(n) epdet(n)  dep(n)    io(n)     msrc(y)   m(y)      munc(n)   mtyp(n)   mdet(n)   mdpsrc(n)         mdpn(n)   mdpix(n)  mdpsc(n)  rem(n)  gid(y)
 #  
 
 class Measure(models.Model):
@@ -13,7 +13,7 @@ class Measure(models.Model):
     day    = models.IntegerField(default=None, null=True, blank=True)
     hou    = models.IntegerField(default=None, null=True, blank=True)
     min    = models.IntegerField(default=None, null=True, blank=True)
-    sec    = models.IntegerField(default=None, null=True, blank=True)
+    sec    = models.FloatField(default=None, null=True, blank=True)
     are    = models.CharField(max_length=255, default='', null=True, blank=True)
     lat    = models.FloatField(null=False, blank=False)  #10
     lon    = models.FloatField(null=False, blank=False)
